Only handle PRs with ID > 500
NOTE: TEMPORARY CHANGE for storybook repo

WHY: pr-log looks at commits with commit messages of the form “Merge
pull request #XXX from …”. When we merged the monorepo, we got a lot of
commits like that which come from repos with different PR numbering. So
I’m filtering them out for the 3.0.0 release.